services/ai-svc/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pgvector==0.3.5
openai==1.54.0
azure-cognitiveservices-speech==1.41.1
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
